+++ CHANGELOG.md
@@ -6,12 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.3.2] - 2023-06-17
+## [1.3.2] - 2023-06-18
 * Enable uint16_t for `BDevices::Keys::KeyType`
 * Add Enable uint16_t for `BDevices::Keys::KeyType::enter`
 * Add hot keys and methods to `BWidgets::Enterable`
 * Add enter by hot key method in `BWidgets::Navigatable`
 * Implement enter by hot key in widgets
+* Fix issue #6 by always updating child widgets in `BWidgets::Box`
 
 
 ## [1.3.1] - 2023-06-17